Start moving response notification email functionality to ProjectFinanceQueries service
# Conflicts:
#	ifs-data-service/src/test/java/org/innovateuk/ifs/notifications/service/FreemarkerNotificationTemplateRendererIntegrationTest.java
#	ifs-data-service/src/test/java/org/innovateuk/ifs/project/security/FinanceCheckServiceSecurityTest.java


Former-commit-id: c4c7931e372c46ce7b5fc879cfd5ed03069a7b01

+++ ifs-data-service/src/main/java/org/innovateuk/ifs/assessment/repository/AssessmentRepository.java
@@ -39,9 +39,9 @@ public interface AssessmentRepository extends ProcessRepository<Assessment>, Pag
 
     List<Assessment> findByActivityStateStateAndTargetCompetitionId(State state, long competitionId);
 
-    long countByActivityStateStateAndTargetCompetitionId(State state, Long competitionId);
+    int countByActivityStateStateAndTargetCompetitionId(State state, Long competitionId);
 
-    long countByActivityStateStateInAndTargetCompetitionId(Collection<State> state, Long competitionId);
+    int countByActivityStateStateInAndTargetCompetitionId(Collection<State> state, Long competitionId);
 
     @Query(value = "SELECT CASE WHEN COUNT(fi.id) = 0" +
             "  THEN 'TRUE'" +
